pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glados"
version = "0.1.0"
description = "GLaDOS persona middleware layer — OpenAI-compatible AI assistant"
requires-python = ">=3.12,<4.0"
dependencies = [
    # HTTP / networking
    "httpx>=0.28.1",
    "requests>=2.32.3",
    "pyyaml>=6.0.2",
    # Data / validation
    "pydantic>=2.10.6",
    "loguru>=0.7.3",
    # NLP / text
    "levenshtein>=0.26.1",
    "rapidfuzz>=3.0.0",      # Stage 3 Phase 1: entity-name fuzzy matching
    "regex>=2024.5.15",
    "jinja2>=3.1.5",
    # Home Assistant WS client (Stage 3 Phase 1)
    "websockets>=13.0",
    # Audio processing (no hardware audio — HA handles playback)
    "soundfile>=0.13.0",
    # ONNX inference — CPU only. Used by glados/ASR/ for speech recognition
    # AND Phase 8.3 semantic entity retrieval (BGE-small-en-v1.5).
    # TTS no longer runs locally; it is delegated to speaches over HTTP.
    # STT will follow in a later change, at which point this can be removed.
    "onnxruntime>=1.16.0",
    # Phase 8.3 — tokenizer for BGE-small. Pure-Rust Hugging Face
    # tokenizers, no PyTorch/transformers dependency chain.
    "tokenizers>=0.20.0",
    # Vision
    "opencv-python-headless>=4.12.0",
    "Pillow>=10.0.0",
    # Memory
    "chromadb>=0.5.0",
    # MCP tool integration
    "mcp>=1.25.0",
    # Auth
    "argon2-cffi>=25.1.0",       # Argon2id password hashing (new default)
    "itsdangerous>=2.2.0",       # signed session cookies
    "bcrypt>=4.0.0",             # legacy hash verification during migration; slated for removal
    # Misc
    "rich>=14.0.0",
    "threadpoolctl>=3.0.0",
    "numba>=0.61.0",
]

[project.optional-dependencies]
# API server
api = ["litestar[standard,structlog]>=2.15.1"]

# Discord integration
discord = ["discord.py>=2.3"]

# Development
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock",
    "ruff",
    "types-requests",
    "types-PyYAML",
]

[project.scripts]
glados = "glados.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["glados*"]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"
exclude = [".venv", "__pycache__", "build", "dist"]

lint.extend-select = [
    "E", "F", "I", "N", "UP", "B", "C4", "RUF",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
